+++ src/objects/loaders/ColladaLoader.js
@@ -23,6 +23,8 @@ THREE.ColladaLoader.prototype = Object.assign(Object.create(THREE.Loader.prototy
 
 		var loader = new THREE.FileLoader(scope.manager);
 		loader.setPath(scope.path);
+		loader.setRequestHeader(scope.requestHeader);
+		loader.setWithCredentials(scope.withCredentials);
 		loader.load(url, function (text) {
 
 			try {
@@ -3891,7 +3893,7 @@ THREE.ColladaLoader.prototype = Object.assign(Object.create(THREE.Loader.prototy
 		// metadata
 
 		var version = collada.getAttribute('version');
-		//console.log('THREE.ColladaLoader: File version', version);
+		console.log('THREE.ColladaLoader: File version', version);
 
 		var asset = parseAsset(getElementsByTagName(collada, 'asset')[0]);
 		var textureLoader = new THREE.TextureLoader(this.manager);
@@ -3961,6 +3963,7 @@ THREE.ColladaLoader.prototype = Object.assign(Object.create(THREE.Loader.prototy
 		setupKinematics();
 
 		var scene = parseScene(getElementsByTagName(collada, 'scene')[0]);
+		scene.animations = animations;
 
 		if (asset.upAxis === 'Z_UP') {
 
@@ -3971,7 +3974,12 @@ THREE.ColladaLoader.prototype = Object.assign(Object.create(THREE.Loader.prototy
 		scene.scale.multiplyScalar(asset.unit);
 
 		return {
-			animations: animations,
+			get animations() {
+
+				console.warn('THREE.ColladaLoader: Please access animations over scene.animations now.');
+				return animations;
+
+			},
 			kinematics: kinematics,
 			library: library,
 			scene: scene