+++ src/objects/loaders/MTLLoader.js
@@ -1,15 +1,7 @@
 const THREE = require('../../three.js');
 
-const MTLLoader = function ( manager ) {
-
-    this.manager = ( manager !== undefined ) ? manager : THREE.DefaultLoadingManager;
-
-};
-
 /**
  * Loads a Wavefront .mtl file specifying materials
- *
- * @author angelxuanchang
  */
 
 THREE.MTLLoader = function (manager) {
@@ -43,6 +35,8 @@ THREE.MTLLoader.prototype = Object.assign(Object.create(THREE.Loader.prototype),
 
 		var loader = new THREE.FileLoader(this.manager);
 		loader.setPath(this.path);
+		loader.setRequestHeader(this.requestHeader);
+		loader.setWithCredentials(this.withCredentials);
 		loader.load(url, function (text) {
 
 			try {
@@ -398,7 +392,7 @@ THREE.MTLLoader.MaterialCreator.prototype = {
 
 					// Diffuse texture map
 
-					setMapForType("map", value);
+					setMapForType('map', value);
 
 					break;
 
@@ -406,7 +400,7 @@ THREE.MTLLoader.MaterialCreator.prototype = {
 
 					// Specular map
 
-					setMapForType("specularMap", value);
+					setMapForType('specularMap', value);
 
 					break;
 
@@ -414,13 +408,13 @@ THREE.MTLLoader.MaterialCreator.prototype = {
 
 					// Emissive map
 
-					setMapForType("emissiveMap", value);
+					setMapForType('emissiveMap', value);
 
 					break;
 
 				case 'norm':
 
-					setMapForType("normalMap", value);
+					setMapForType('normalMap', value);
 
 					break;
 
@@ -429,7 +423,7 @@ THREE.MTLLoader.MaterialCreator.prototype = {
 
 					// Bump texture map
 
-					setMapForType("bumpMap", value);
+					setMapForType('bumpMap', value);
 
 					break;
 
@@ -437,7 +431,7 @@ THREE.MTLLoader.MaterialCreator.prototype = {
 
 					// Alpha map
 
-					setMapForType("alphaMap", value);
+					setMapForType('alphaMap', value);
 					params.transparent = true;
 
 					break;